requirements.txt
fastapi==0.110.0
pymongo==3.12.0
numpy==1.21.5
pillow==9.2.0

uvicorn==0.28.0
tensorflow==2.11.0
